UPDATE CODE REFACTO:
Code cleaning and separating into different functions
