+++ front/app/src/Views/Chat/ChatLayout.tsx
@@ -1,28 +1,19 @@
+/* eslint-disable no-alert */
 /* eslint-disable curly */
 /* eslint-disable max-statements */
 /* eslint-disable max-len */
 /* eslint-disable max-lines-per-function */
 
 import {
-	AppBar,
-	Avatar,
 	Box,
 	Button,
-	Card,
-	CardContent,
-	CardMedia,
 	Divider,
 	Fab,
 	Grid,
-	IconButton,
 	List,
 	ListItem,
-	ListItemIcon,
 	ListItemText,
-	ListItemTextProps,
-	MenuItem,
 	Paper,
-	Select,
 	Tab,
 	Tabs,
 	TextField,
@@ -35,79 +26,35 @@ import {
 	DialogTitle
 } from "@mui/material";
 
+import MessageItem from "./components/MessageItem";
+import CurrentlyTalkingFriend from "./components/CurrentlyTalkingFriend";
+import FriendItem from "./components/FriendItem";
+
 const URL = "http://localhost:3000";
 import SendIcon from "@mui/icons-material/Send";
-import PlayArrowIcon from "@mui/icons-material/PlayArrow";
 import MenuBar from "../../Component/MenuBar/MenuBar";
 import { useTheme } from "@emotion/react";
-import { CSSProperties, useState, useEffect, useRef } from "react";
+import
+{
+	CSSProperties,
+	useState,
+	useEffect,
+	useRef
+}	from "react";
 import { motion } from "framer-motion";
-import { Outlet } from "react-router-dom";
-import { Socket, io } from "socket.io-client";
-import { socket, connect, disconnect } from "./socket.service";
+import { io } from "socket.io-client";
 
-// please use vector this one is just for testing card
-import pong from "./assets/pong.jpeg";
-import { useAppDispatch, useAppSelector } from "../../Redux/hooks/redux-hooks";
-import { addMessage, setActiveConversationId, setCurrentChannel, setChatConnected, setChatUsers, setMessageRoom } from "../../Redux/store/controllerAction";
+import {
+	useAppDispatch,
+	useAppSelector
+} from "../../Redux/hooks/redux-hooks";
+import {
+	setActiveConversationId,
+	setCurrentChannel,
+	setChatUsers,
+	setMessageRoom
+}	from "../../Redux/store/controllerAction";
 import { MessageRoomModel } from "../../Redux/models/redux-models";
-import { DataArraySharp } from "@mui/icons-material";
-import { current } from "@reduxjs/toolkit";
-
-// invite
-const InvitationCard = () => {
-	const theme = useTheme();
-	return (
-		<Card sx={{
-			display: "flex",
-			justifyContent: "flex-start",
-			backgroundColor: theme.palette.background.default,
-			width: "40%",
-		}}>
-			<Box sx={
-				{
-					display: "flex",
-					flexDirection: "column"
-				}
-			}>
-				<CardContent sx={{ flex: "1 0 auto" }}>
-					<Typography component="div" variant="h5">
-						Play Pong with me
-					</Typography>
-					<Typography variant="subtitle1" color="text.secondary" component="div">
-						A slot is reserved between John Wick and you
-					</Typography>
-				</CardContent>
-				<Box sx={
-					{
-						display: "flex",
-						pl: 1,
-						pb: 1
-					}}
-				>
-					{/* <IconButton aria-label="previous">
-						{theme.direction === "rtl" ? <SkipNextIcon /> : <SkipPreviousIcon />}
-					</IconButton> */}
-					<IconButton aria-label="play/pause">
-						<PlayArrowIcon sx={{
-							height: 38,
-							width: 38
-						}} />
-					</IconButton>
-					{/* <IconButton aria-label="next">
-						{theme.direction === "rtl" ? <SkipPreviousIcon /> : <SkipNextIcon />}
-					</IconButton> */}
-				</Box>
-			</Box>
-			<CardMedia
-				component="img"
-				sx={{ width: 200 }}
-				image={pong}
-				alt="Live from space album cover"
-			/>
-		</Card>
-	);
-};
 
 type MessageModel =
 {
@@ -132,23 +79,11 @@ interface TabPanelProps {
 	area?: boolean;
 }
 
-const ChannelsList = () => {
-	return (
-		<>
-			channel list here, you can follow FriendsList component
-			<br />
-			<Button variant="contained" color="success"> onClick={createNewChannel}
-				NEW
-			</Button>
-		</>
-	);
-};
-
-const TabPanel = (props: TabPanelProps) => {
+const TabPanel = (props: TabPanelProps) =>
+{
 	const { children, value, index, ...other } = props;
 	let animationSettings;
 
-	// this is a switch I added to change animation type on left | right view
 	if (props.area === false)
 		animationSettings = {
 			type: "spring",
@@ -196,78 +131,22 @@ const TabPanel = (props: TabPanelProps) => {
 	);
 };
 
-// can be used to display the list of friends
-const CurrentlyTalkingFriend = () => {
-	const currentlyTalkingFriendDataFake = {
-		name: "John Wick",
-		avatar: "https://material-ui.com/static/images/avatar/1.jpg",
-	};
-
-	return (
-		<List>
-			<ListItem
-				button
-				key={currentlyTalkingFriendDataFake.name}
-			>
-				<ListItemIcon>
-					<Avatar
-						alt={currentlyTalkingFriendDataFake.name}
-						src={currentlyTalkingFriendDataFake.avatar}
-					/>
-				</ListItemIcon>
-				<ListItemText primary={currentlyTalkingFriendDataFake.name}>
-				</ListItemText>
-			</ListItem>
-		</List>
-	);
-};
-
-type FriendItemProps = {
-	name: string;
-	avatar?: string;
-	online?: boolean;
-	key?: number;
-};
-const FriendItem = (props: FriendItemProps) => {
-	const status = props.online ? "online 💚" : "🔴";
-
-	return (
-		<ListItem
-			button
-			key={props.key}
-		>
-			<ListItemIcon>
-				<Avatar
-					alt={props.name}
-					src={props.avatar}
-				/>
-			</ListItemIcon>
-			<ListItemText primary={props.name}>
-				{props.name}
-			</ListItemText>
-			<ListItemText
-				secondary={status}
-				sx={{ align: "right" }}
-			>
-			</ListItemText>
-		</ListItem>
-	);
-};
-
 type FriendsListProps = {
 	arrayListUsers: []
 };
 
-const FriendsList = (props: FriendsListProps) => {
-	const socketTest = useRef<SocketIOClient.Socket | null>(null);
-	let friendList: any[];
-	const dispatch = useAppDispatch();
-	const users = useAppSelector((state) =>
+const FriendsList = (props: FriendsListProps) =>
+{
+	const	socketTest = useRef<SocketIOClient.Socket | null>(null);
+	let		friendList: any[];
+	const	dispatch = useAppDispatch();
+	const	users = useAppSelector((state) =>
 	{
 		return (state.controller.user.chat.users);
 	});
 
-	const sendMsg = (id: string) => {
+	const sendMsg = (id: string) =>
+	{
 		const action = {
 			type: "sending-message"
 		};
@@ -275,7 +154,8 @@ const FriendsList = (props: FriendsListProps) => {
 		// socketTest.current?.emit("send-message", action);
 	};
 
-	const displayConversationWindow = (id: string) => {
+	const displayConversationWindow = (id: string) =>
+	{
 		const action = {
 			type: "display-conversation",
 			payload:
@@ -306,10 +186,12 @@ const FriendsList = (props: FriendsListProps) => {
 			<Divider />
 			<List>
 				{
-					users.map((elem, index) => {
+					users.map((elem, index) =>
+					{
 						return (
 							<>
-								<div onClick={() => {
+								<div onClick={() =>
+								{
 									displayConversationWindow(elem.id);
 									dispatch(setActiveConversationId(elem.id));
 								}}>
@@ -329,86 +211,9 @@ const FriendsList = (props: FriendsListProps) => {
 	);
 };
 
-// the next line interface is here to remove the typeScript errordeclare module '@mui/material/ListItemText'
-declare module "@mui/material/ListItemText"
+
+const MessagesArea = (text: any) =>
 {
-	interface ListItemTextProps {
-		align?: "left" | "center" | "right";
-	}
-}
-
-type MessageItemProps = {
-	key?: number,
-	align?: ListItemTextProps,
-	sender: "me" | "other" | "server",
-	message: string,
-	date: string
-};
-
-const MessageItem = (props: MessageItemProps) => {
-	let align: "right" | "center" | "left" | undefined;
-
-	switch (props.sender) {
-		case "me":
-			align = "right";
-			break;
-		case "other":
-			align = "left";
-			break;
-		case "server":
-			align = "center";
-			break;
-		default:
-			align = "center";
-			break;
-	}
-	if (props.message === "!play pong" && props.sender === "server")
-		return (
-			<ListItem key={props.key}>
-				<Grid container>
-					<Grid item xs={12}>
-						<ListItemText
-							align={align}
-							color="primary"
-						>
-							<InvitationCard />
-						</ListItemText>
-					</Grid>
-					<Grid item xs={12}>
-						<ListItemText
-							align={align}
-							secondary={props.date}
-						>
-						</ListItemText>
-					</Grid>
-				</Grid>
-			</ListItem>
-		);
-	else
-		return (
-			<ListItem key={props.key}>
-				<Grid container>
-					<Grid item xs={12}>
-						<ListItemText
-							align={align}
-							color="primary"
-							primary={props.message}
-						>
-						</ListItemText>
-					</Grid>
-					<Grid item xs={12}>
-						<ListItemText
-							align={align}
-							secondary={props.date}
-						>
-						</ListItemText>
-					</Grid>
-				</Grid>
-			</ListItem>
-		);
-};
-
-const MessagesArea = (text: any) => {
 	let displayMessageArray;
 
 	displayMessageArray = [
@@ -443,19 +248,23 @@ const MessagesArea = (text: any) => {
 			},
 		];
 	}
-	else {
+	else
+	{
 		const msgRoom = users[userActiveIndex].msgRoom;
 		let i;
 
 		i = 0;
-		while (msgRoom.length) {
-			if (msgRoom[i].id === activeId) {
+		while (msgRoom.length)
+		{
+			if (msgRoom[i].id === activeId)
+			{
 				displayMessageArray = msgRoom[i].content;
 				break;
 			}
 			i++;
 		}
-		if (i === msgRoom.length) {
+		if (i === msgRoom.length)
+		{
 			displayMessageArray = [
 				{
 					sender: "server",
@@ -475,7 +284,8 @@ const MessagesArea = (text: any) => {
 				}}
 			>
 				{
-					displayMessageArray.map((elem, key) => {
+					displayMessageArray.map((elem, key) =>
+					{
 						return (
 							<MessageItem
 								key={key}
@@ -490,50 +300,6 @@ const MessagesArea = (text: any) => {
 		</>
 	);
 };
-
-// const SendingArea = () =>
-// {
-// 	const
-// 	[
-// 		text,
-// 		setText
-// 	] = useState("");
-// 	const handleTextChange = (e: any) =>
-// 	{
-// 		setText(e.target.value);
-// 	};
-
-// 	const handleSendClick = () =>
-// 	{
-// 		console.log("Text typed:", text);
-// 		const	socketRef = useRef<SocketIOClient.Socket | null>(null);
-// 		const action = {
-// 			type: "test",
-// 			payload: "Hello-world"
-// 		};
-// 		socketRef.current.emit("info", action);
-// 		MessagesArea(text);
-// 	};
-
-// 	return (
-// 		<>
-// 		<Grid item xs={11}>
-// 			<TextField
-// 			id="outlined-basic-email"
-// 			label="Type Something"
-// 			fullWidth
-// 			value={text}
-// 			onChange={handleTextChange}
-// 			/>
-// 		</Grid>
-// 		<Grid xs={1} sx={{ alignItems: "right" }}>
-// 			<Fab color="primary" aria-label="add" onClick={handleSendClick}>
-// 			<SendIcon />
-// 			</Fab>
-// 		</Grid>
-// 		</>
-// 	);
-// };
 
 const a11yProps = (index: any) =>
 {
@@ -621,11 +387,13 @@ const	ChatLayout = () =>
 		setChannelMembers
 	] = useState<MembersModel[]>([]);
 
-	const handleClickOpen = () => {
+	const handleClickOpen = () =>
+	{
 		setOpen(true);
 	};
 
-	const handleClose = () => {
+	const handleClose = () =>
+	{
 		setChannelName("");
 		setSelectedMode("");
 		setChanPassword("");
@@ -754,7 +522,7 @@ const	ChatLayout = () =>
 			{
 				if (data.payload.message === "")
 					setChannels(data.payload.chanMap);
-				elsethis
+				else
 					alert(data.payload.message);
 			}
 
@@ -765,7 +533,7 @@ const	ChatLayout = () =>
 				else
 				{
 					setChanMessages([]);
-					alert ("Successfully joined channel " + data.payload.chanName + "!");
+					alert("Successfully joined channel " + data.payload.chanName + "!");
 				}
 			}
 
@@ -794,7 +562,8 @@ const	ChatLayout = () =>
 			// setArrayListUser(data.payload.arrayListUser);
 		};
 
-		const sendMessageToUser = (data: any) => {
+		const sendMessageToUser = (data: any) =>
+		{
 			const msgRoom: MessageRoomModel[] = [
 				{
 					id: data.payload.msgRoom.id,
@@ -906,7 +675,8 @@ const	ChatLayout = () =>
 		setValue(index);
 	};
 
-	const refreshListUser = () => {
+	const refreshListUser = () =>
+	{
 		// if (chatConnected === false)
 		// {
 		const action = {
@@ -995,11 +765,13 @@ const	ChatLayout = () =>
 		setChannelAction
 	] = useState("");
 
-	const handleDialogOpen = () => {
+	const handleDialogOpen = () =>
+	{
 		setIsDialogOpen(true);
 	};
 
-	const handleDialogClose = () => {
+	const handleDialogClose = () =>
+	{
 		setIsDialogOpen(false);
 	};
 
@@ -1122,7 +894,8 @@ const	ChatLayout = () =>
 										type="text"
 										placeholder="Channel name"
 										value={channelName}
-										onChange={(e) => {
+										onChange={(e) =>
+										{
 											const inputValue = e.target.value;
 											if (inputValue.length <= 8)
 												setChannelName(inputValue);
@@ -1138,7 +911,10 @@ const	ChatLayout = () =>
 										name="answerOption"
 										value="public"
 										checked={selectedMode === "public"}
-										onChange={() => setSelectedMode("public")}
+										onChange={() =>
+										{
+											setSelectedMode("public");
+										}}
 										/>
 										<label htmlFor="option1">Public</label>
 									</div>
@@ -1149,7 +925,10 @@ const	ChatLayout = () =>
 										name="answerOption"
 										value="protected"
 										checked={selectedMode === "protected"}
-										onChange={() => setSelectedMode('protected')}
+										onChange={() =>
+										{
+											setSelectedMode("protected");
+										}}
 										/>
 										<label htmlFor="option2">Protected</label>
 									</div>
@@ -1160,7 +939,10 @@ const	ChatLayout = () =>
 										name="answerOption"
 										value="private"
 										checked={selectedMode === "private"}
-										onChange={() => setSelectedMode("private")}
+										onChange={() =>
+										{
+											setSelectedMode("private");
+										}}
 										/>
 										<label htmlFor="option3">Private</label>
 									</div>
@@ -1168,7 +950,10 @@ const	ChatLayout = () =>
 										type="text"
 										placeholder="Password (if protected)"
 										value={chanPassword}
-										onChange={(e) => setChanPassword(e.target.value)}
+										onChange={(e) =>
+										{
+											setChanPassword(e.target.value);
+										}}
 									/>
 								</DialogContent>
 								<DialogActions>
@@ -1181,109 +966,128 @@ const	ChatLayout = () =>
 								</DialogActions>
 							</Dialog>
 							<List>
-								{channels.map((channel: any) => {return (
-								<ListItem style={listItemStyle} key={channel.id}>
-									<ListItemText
-										style={
-											channel.name === currentChannel
-											? { color: "red" }
-											: listItemTextStyle
-										}
-										primary={channel.name}
-										onClick={() =>
-										{
-											return (goToChannel(channel.name));
-										}}
-									/>
-									<Button onClick={handleDialogOpen}>Options</Button>
-									{/* Dialog component */}
-									<Dialog open={isDialogOpen} onClose={handleDialogClose}>
-										<DialogTitle>Choose an Action</DialogTitle>
-										<DialogContent>
-											<Button onClick={() =>
-											{
-												return handleJoinButtonClick(channel.mode, channel.name);
-											}}>
-												Join
-											</Button>
-											<Button onClick={() =>
-											{
-												return handleLeaveButtonClick(channel.id, channel.name);
-											}}>
-												Leave
-											</Button>
-											<Button onClick={() =>
-											{
-												return handleRemoveButtonClick(channel.id, channel.name);
-											}}>
-												Remove
-											</Button>
-											<Button onClick={() =>
-											{
-												return handleMembersClickOpen(channel.name);
-											}}>Members</Button>
-											<Dialog open={membersOpen} onClose={handleMembersClose} maxWidth="sm" fullWidth>
-												<DialogTitle>Channel Members</DialogTitle>
-												<DialogContent>
-												<ul>
-													{channelMembers.map((member) =>
+								{
+									channels.map((channel: any) =>
+									{
+										return (
+											<ListItem style={listItemStyle} key={channel.id}>
+												<ListItemText
+													style={
+														channel.name === currentChannel
+														? { color: "red" }
+														: listItemTextStyle
+													}
+													primary={channel.name}
+													onClick={() =>
 													{
-														return (
-															<li key={member.id}>{member.name}</li>);
-													})}
-												</ul>
-												</DialogContent>
-												<DialogActions>
-												<Button onClick={handleMembersClose} color="primary">
-													Close
+														return (goToChannel(channel.name));
+													}}
+												/>
+												<Button onClick={handleDialogOpen}>
+													Options
 												</Button>
-												</DialogActions>
-											</Dialog>
-										</DialogContent>
-										<DialogActions>
-										<Button onClick={handleDialogClose} color="primary">
-											Cancel
-										</Button>
-										</DialogActions>
-									</Dialog>
-									<Dialog open={openPasswordDialog} onClose={() =>
-										{
-											setOpenPasswordDialog(false);
-										}}>
-										<DialogTitle>
-											Enter Password
-										</DialogTitle>
-										<DialogContent>
-											<TextField
-												label="Password"
-												type="password"
-												fullWidth
-												variant="outlined"
-												value={userPassword}
-												onChange={(e) =>
-												{
-													setUserPassword(e.target.value);
-												}}
-											/>
-										</DialogContent>
-										<DialogActions>
-											<Button onClick={() => setOpenPasswordDialog(false)} color="primary">
-												Cancel
-											</Button>
-											<Button
-												onClick={() =>
-												{
-													const password = userPassword;
-													handlePasswordSubmit(userPassword);
-												}}
-												color="primary"
+												<Dialog open={isDialogOpen} onClose={handleDialogClose}>
+													<DialogTitle>
+														Choose an Action
+													</DialogTitle>
+													<DialogContent>
+														<Button onClick={() =>
+														{
+															return handleJoinButtonClick(channel.mode, channel.name);
+														}}>
+															Join
+														</Button>
+														<Button onClick={() =>
+														{
+															return handleLeaveButtonClick(channel.id, channel.name);
+														}}>
+															Leave
+														</Button>
+														<Button onClick={() =>
+														{
+															return handleRemoveButtonClick(channel.id, channel.name);
+														}}>
+															Remove
+														</Button>
+														<Button onClick={() =>
+														{
+															return handleMembersClickOpen(channel.name);
+														}}>Members</Button>
+														<Dialog open={membersOpen} onClose={handleMembersClose} maxWidth="sm" fullWidth>
+															<DialogTitle>
+																Channel Members
+															</DialogTitle>
+															<DialogContent>
+																<ul>
+																	{
+																		channelMembers.map((member) =>
+																		{
+																			return (<li key={member.id}>{member.name}</li>);
+																		})
+																	}
+																</ul>
+															</DialogContent>
+															<DialogActions>
+															<Button onClick={handleMembersClose} color="primary">
+																Close
+															</Button>
+															</DialogActions>
+														</Dialog>
+													</DialogContent>
+													<DialogActions>
+													<Button onClick={handleDialogClose} color="primary">
+														Cancel
+													</Button>
+													</DialogActions>
+												</Dialog>
+												<Dialog
+													open={openPasswordDialog}
+													onClose={() =>
+													{
+														setOpenPasswordDialog(false);
+													}}
 												>
-												Submit
-											</Button>
-										</DialogActions>
-									</Dialog>
-								</ListItem>
-								)})}
+													<DialogTitle>
+														Enter Password
+													</DialogTitle>
+													<DialogContent>
+														<TextField
+															label="Password"
+															type="password"
+															fullWidth
+															variant="outlined"
+															value={userPassword}
+															onChange={(e) =>
+															{
+																setUserPassword(e.target.value);
+															}}
+														/>
+													</DialogContent>
+													<DialogActions>
+														<Button
+															onClick={() =>
+															{
+																setOpenPasswordDialog(false);
+															}}
+															color="primary"
+														>
+															Cancel
+														</Button>
+														<Button
+															onClick={() =>
+															{
+																handlePasswordSubmit(userPassword);
+															}}
+															color="primary"
+															>
+															Submit
+														</Button>
+													</DialogActions>
+												</Dialog>
+											</ListItem>
+										);
+									})
+								}
 							</List>
 						</div>
 					</TabPanel>
@@ -1297,9 +1101,7 @@ const	ChatLayout = () =>
 						<FriendsList arrayListUsers={arrayListUser} />
 					</TabPanel>
 				</Grid>
-				{/* left side of the screen  */}
 				<Grid item xs={9}>
-					{/* when value == 0  */}
 					<TabPanel
 						area={true}
 						value={value}
